Move get methods to correct forms
