+++ Kitodo/src/main/java/de/sub/goobi/forms/ProjekteForm.java
@@ -800,18 +800,4 @@ public class ProjekteForm extends BasisForm {
     public int getItemId() {
         return this.itemId;
     }
-
-    /**
-     * Return list of projects
-     *
-     * @return list of projects
-     */
-    public List<ProjectDTO> getProjects() {
-        try {
-            return serviceManager.getProjectService().findAll();
-        } catch (DataException e) {
-            logger.error("Unable to load projects: " + e.getMessage());
-            return new LinkedList<>();
-        }
-    }
 }
